Links aangepast controller

+++ song/src/main/java/com/projectjava/song/entity/Song.java
@@ -5,11 +5,13 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
+
 @Data
 @Document(collection = "Song")
 public class Song {
     @Id
-    private int id;
+    private BigInteger id;
     private String title;
     private String artist;
     private String genre;
